MPI codee works

+++ monte_carlo/main_mpi.c
@@ -4,7 +4,7 @@
 #include <unistd.h>
 #include <cmath>
 #include <sstream>
-#include<iomanip>
+#include <iomanip>
 #include <fstream>
 #include <mpi.h>
 /**************************************************/
@@ -37,7 +37,7 @@ int main(int argc, char *argv[]){
     //
     mpi_err = MPI_Init(0x0, 0x0);
     mpi_err =  MPI_Comm_rank(MPI_COMM_WORLD, &mpi_rank);
-    seed_v = 7*3*11*(mpi_rank+1)*rand();
+    seed_v = (uint32_t) 7*3*11*(mpi_rank+1)*rand();
     init_rng(seed_v);
     //
     outfolder = ZeroPadNumber(mpi_rank);
